This work is derived from https://github.com/JuliaStats/TimeSeries.jl/pull/321#discussion_r128827266

This PR make generic function do `broadcast` call correctly.

e.g.
```julia
julia> using MarketData

julia> f(a, b, c, d, e) = a * b + c - d + e 

julia> f.(1, 1, cl, ohlc, op[1:10])  # this was borken before this PR
10x4 TimeSeries.TimeArray{Float64,2,Date,Array{Float64,2}} 2000-01-03 to 2000-01-14 

             Close_Open_Open  Close_High_Open  Close_Low_Open  Close_Close_Open 
2000-01-03 | 112.94           105.32           116.13          105.88     
2000-01-04 | 103.5            101.13           110.56          109.25
2000-01-05 | 105.0            98.19            105.75          104.75      
2000-01-06 | 96.0             95.12            107.12          107.12
⋮
2000-01-11 | 93.75            90.31            99.19           96.94
2000-01-12 | 88.19            87.69            96.69           96.0
2000-01-13 | 97.75            93.48            99.73           95.48  
2000-01-14 | 101.44           99.19            102.06          101.0
```
---

### Performance

The performance of current implementation is quite *poor* and the memory requirement is larger :(.
(I'm going to read more about performance tips. Please advise)

A test case for `log.(2, cl .+ 1)` which the works fine before.
Before:
```julia
julia> @benchmark log.(2, $cl .+ 1) 
BenchmarkTools.Trial:               
  memory estimate:  38.28 KiB       
  allocs estimate:  1269            
  --------------                    
  minimum time:     111.731 μs (0.00% GC)                                
  median time:      114.278 μs (0.00% GC)                                
  mean time:        118.402 μs (2.94% GC)                                
  maximum time:     2.330 ms (91.95% GC)                                 
  --------------                    
  samples:          10000           
  evals/sample:     1
```

After:
```julia
julia> @benchmark log.(2, $cl .+ 1) 
BenchmarkTools.Trial:               
  memory estimate:  85.19 KiB       
  allocs estimate:  2554            
  --------------                    
  minimum time:     218.412 μs (0.00% GC)                                
  median time:      224.006 μs (0.00% GC)                                
  mean time:        238.519 μs (4.23% GC)                                
  maximum time:     3.578 ms (87.72% GC)                                 
  --------------                    
  samples:          10000           
  evals/sample:     1
```

----

### TODO
- [x] add test cases
- [x] improve performance