@iblis17 2017-09-03
current patch set:
```julia
julia> @benchmark log.(2, $cl .+ 1)                                                                                                               
BenchmarkTools.Trial:               
  memory estimate:  41.47 KiB       
  allocs estimate:  1251            
  --------------                    
  minimum time:     93.868 μs (0.00% GC)                                                                                                          
  median time:      96.204 μs (0.00% GC)                                 
  mean time:        100.758 μs (3.53% GC)                                                                                                         
  maximum time:     2.244 ms (92.37% GC)                                                                                                          
  --------------                    
  samples:          10000           
  evals/sample:     1               

shell> git log | head                                                                                                                             
commit da49c48a78f633d5e56c3659d38bde1132212926                                                                                                   
Author: Iblis Lin <iblis@hs.ntnu.edu.tw>                                                                                                          
Date:   Sun Sep 3 19:19:13 2017 +0800                                                                                                             
                                                                                                                                                  
    broadcast: wip: rewrite broadcast_c in generated func                                                                                         
```