@iblis17 2017-10-23
I'm going to merge this PR tmr, if no objection